% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateEB.R
\name{calibrateEB}
\alias{calibrateEB}
\alias{gfit}
\alias{gbayes}
\title{Empirical Bayes calibration of noisy variance estimates}
\usage{
calibrateEB(vars, sigma2)

gfit(X, sigma, p = 2, nbin = 1000, unif.fraction = 0.1)

gbayes(x0, g.est, sigma)
}
\arguments{
\item{vars}{list of variance estimates}

\item{sigma2}{estimate of the Monte Carlo noise in vars}

\item{X}{a vector of observations}

\item{sigma}{noise estimate}

\item{p}{tuning parameter -- number of parameters used to fit G}

\item{nbin}{tuning parameter -- number of bins used for discrete approximation}

\item{unif.fraction}{tuning parameter -- fraction of G modeled as "slab"}

\item{x0}{an obsevation}

\item{g.est}{a prior density, as returned by gfit}

\item{sigma}{noise estimate}
}
\value{
calibrated variance estimates

posterior density estimate g

posterior estimate E[mu | x0]
}
\description{
Shamelessly lifted from Stefan's deprecated randomForestCI-package on https://github.com/swager/randomForestCI

mu ~ G, X ~ N(mu, sigma^2)

Bayes posterior estimation with Gaussian noise
}
\section{References}{

For more details about "g-estimation", see: B Efron. Two modeling strategies for
empirical Bayes estimation. Stat. Sci., 29(2): 285–301, 2014.
}

\author{
Stefan Wager
}
