% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.ncdf.R
\name{extract.ncdf}
\alias{extract.ncdf}
\title{Extract data and attributes (e.g. model structure) from netCDF (.nc) file}
\usage{
extract.ncdf(ncfile)
}
\arguments{
\item{ncfile}{netCDF input file (character with extension .nc).}
}
\value{
A list containing [[1]] a list of attributes (information and instruction), and [[2]] a \code{data.frame} with the simulated data.
}
\description{
This function allows to readily extract data, simulated with simSAC::simData, and general information (attributes) from the produced netCDF (.nc) file.
}
\examples{
\dontrun{
# simulate data for smooth landscape, normally distributed y and no SAC
simData("110") 
# extract attributes and data
SGR <- extract.ncdf(paste0(getwd(),"/dataset110.nc")) 
SGR[[1]] # attributes
head(SGR[[2]]) # data

library(lattice)
levelplot(y~Lon+Lat,data=SGR[[2]]) # plot response varibale on Longitude-Latitude grid
}
}
\seealso{
\code{\link{simData}}
}
\author{
Severin Hauenstein <severin.hauenstein@biom.uni-freiburg.de>
}
