% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.to.num.R
\name{geo.to.num}
\alias{geo.to.num}
\title{Transform geographic coordinates to numeric values}
\usage{
geo.to.num(geo)
}
\arguments{
\item{geo}{Geographic coordinate (Character).}
}
\value{
Returns the coordinates as numeric values. To be used for the creation of (\code{\link[raster]{extent}}) objects.
}
\description{
This function transforms geographic coordinates, such as \code{"5N24E"} or \code{"7S27E"} to numeric values, such as \code{c(5, 24)} or \code{c(-7, 27)}.
}
\examples{
\dontrun{
# Transform "5N24E" or "7S27E" to numeric values
coords <- c("5N24E", "7S27E")
num.coords <- sapply(1:2, function(x) geo.to.num(coords[x]))
}
}
\seealso{
Function is used in \code{\link{simData}}.
}
\author{
Severin Hauenstein <severin.hauenstein@biom.uni-freiburg.de>
}
