% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.asking.R
\name{keep.asking}
\alias{keep.asking}
\title{Wait for specific keypress}
\usage{
keep.asking(Q, A = c("y", "n"),
  Reminder = "Please enter: y for yes, n for no.")
}
\arguments{
\item{Q}{Question to be asked (Character).}

\item{A}{Vector of possible answers. Defaults to \code{c("y","n")}}

\item{Reminder}{If wrong answer was entered, reminds of the possible answers. Defaults to \code{"Please enter: y for yes, n for no."}}
}
\value{
Returns the answer entered by the user (Character).
}
\description{
This function allows to ask a question and prompts to a specific answer.
}
\examples{
# check with user if data really should be downloaded
\dontrun{
check.download <- keep.asking(Q = "Do you want to download data from http://www.worldclim.org?")
y
}
}
\seealso{
Function is used in \code{\link{simData}}.
}
\author{
Severin Hauenstein <severin.hauenstein@biom.uni-freiburg.de>
}
