% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polis.R
\docType{data}
\name{polis}
\alias{polis}
\title{Presence/absence of Uta-lizards on Gulf of California islands}
\format{An object of class \code{data.frame} with 20 rows and 4 columns.}
\usage{
data(polis)
}
\description{
An example of a small data set for binary response
}
\details{
Data are from Polis et al. (1998, table 1), provided by Quinn & Keough (2002). They report the presence or absence of lizards of the genus \emph{Uta}) on 20 different islands in the Gulf of California, alongside with the parameter-to-area ratio.
}
\references{
Polis, G.A., Hurd, S.D., Jackson, C.T., Sanchez-Piñero, F., Ecology, S., Mar, N. & Jackson, T. (2014) Multifactor Population Limitation : Variable Spatial and Temporal Control of Spiders on Gulf of California Islands. \emph{Ecology} \bold{79}, 490-–502

Quinn, G.P. & Keough, M.J. (2002) \emph{Experimental Design and Data Analysis for Biologists}. Cambridge University Press, Cambridge, UK.
}
\author{
Carsten F. Dormann \email{carsten.dormann@biom.uni-freiburg.de}
}
\keyword{data}
