% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeit.R
\name{rangeit}
\alias{rangeit}
\title{Range data between 0 and 1}
\usage{
rangeit(x, na.rm = TRUE)
}
\arguments{
\item{x}{Data vector to be ranged.}

\item{na.rm}{Logical; should NA-values be ignored when computing min and max? Defaults to TRUE:}
}
\value{
a vector of same length as x
}
\description{
Re-scale the data to lie between 0 and 1
}
\details{
A trivial function to re-scale data by subtracting the min and dividing by the difference between max and min. This function does not change the shape of the distribution.
}
\examples{
blubb <- rnorm(100, sd=15)
par(mfrow=c(1,2))
hist(blubb)
hist(rangeit(blubb))
par(mfrow=c(1,1))
}
\seealso{
\code{scale}
}
\author{
Carsten F. Dormann <carsten.dormann@biom.uni-freiburg.de>
}
