% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinitesimalJackknife.R
\name{rangerInfJack}
\alias{rangerInfJack}
\alias{rangerInfJackMulticlass}
\alias{rInfJack}
\title{Deprecated: Compute variance of estimate from a "ranger" model}
\usage{
rangerInfJack(rf, newdata, calibrate = TRUE)

rangerInfJackMulticlass(rf, newdata, calibrate = TRUE)

rInfJack(pred, inbag, calibrate = TRUE, used.trees = NULL)
}
\arguments{
\item{rf}{A ranger object, fitted using the function ranger in \pkg{ranger}, as a faster alternative to \pkg{randomForest}.}

\item{calibrate}{whether to apply calibration to mitigate Monte Carlo noise
warning: if calibrate = FALSE, some variance estimates may be negative
         due to Monte Carlo effects if the number of trees in rf is too small}

\item{pred}{A nrow(newdata) by no. of trees matrix which contains numeric predictions
from a random forest trained with trees grown on bootstrap samples of the training data}

\item{inbag}{A number of obs. in the training data by no. of trees matrix giving the
number of times the ith observation in the training data appeared in the bootstrap sample for the jth tree.}

\item{used.trees}{set of trees to use for variance estimation; uses all tress if NULL}
}
\value{
For regressions, a two-column matrix is returned, with predictions in the first column and estimates of prediction variance in the second. For classification, a 2*k-column matrix is returned, with the probability estimates for the k classes in the first k columns, and their variance estimates in the next k columns.
}
\description{
Computes variances for a prediction from a ranger-model, using the infinitesimal jackknife procedure
}
\details{
This function is a ranger-adaptation of the package \pkg{randomForestCI} of Wager et al. (2014). Their original can be found on github: \url{ https://github.com/swager/randomForestCI/}. All that the three functions below do differently is to allow using the ranger model object (which is computed much faster than the classical randomForest in \pkg{randomForest}). The functions are named differently to avoid confusion. Instead, use the same functions as provided in \pkg{ranger} now!

For classification problems, use \code{rangerInfJackMulticlass}; for regression problems use \code{rangerInfJack}. \code{rInfJack} is the internal workhorse called by the other two functions. It is probably of little relevance for the typical user.
}
\note{
The warning "No calibration with n<= 20" indicates that the variance estimates were not re-calibrated across the predictions. Please read the original paper to understand what that means (Wager et al. 2014).
}
\examples{
\dontrun{ # because the packages with the data are not in the dependencies of FReibier
# a binary classification example using the Titanic-data:
library(effects)
franger <- ranger(survived ~ age+sex+passengerClass, data=na.omit(TitanicSurvival), keep.inbag=T, replace=T)
rangerInfJackMulticlass(franger, newdata=na.omit(TitanicSurvival[1:300,]), calibrate=T)

# a regression example using the alfalfa data:
library(faraway)
fr <- ranger(yield ~ shade + irrigation + inoculum, data=alfalfa, keep.inbag=T, replace=T)
rangerInfJack(fr, newdata=alfalfa[1:2,], calibrate=F)
}
}
\references{
Wager, Stefan, Trevor Hastie, and Bradley Efron. 2014. Confidence intervals for random forests: The jackknife and the infinitesimal jackknife. \emph{The Journal of Machine Learning Research} \bold{15} (1), 1625--1651.
}
\author{
Carsten F. Dormann <carsten.dormann@biom.uni-freiburg.de>
}
