% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiders8.R
\docType{data}
\name{spiders8}
\alias{spiders8}
\title{Spiders caught in 16 traps in Europe}
\format{An object of class \code{matrix} with 16 rows and 105 columns.}
\usage{
data(spiders8)
}
\description{
An example dataset of a community matrix (with traps instead of sites) and integer values (i.e. number of individuals per species)
}
\details{
These are real data from the Greenveins project, representing one of the sites (site 8, hence the name). For details see reference (data are also included in the PREDICTs data base). Columns are species (abbreviated as first two letters of the genus in capitals, followed by the epithet in lower case). Rows are the 16 "trap sets" (ts). Numbers are individuals per species.
}
\author{
Carsten F. Dormann \email{carsten.dormann@biom.uni-freiburg.de}
}
\keyword{data}
